requirements.txt
Cython==0.29.19
llvmlite==0.32.1
numba==0.49.1
numpy==1.18.4
Pillow==7.1.2
pygame==1.9.6
scipy==1.4.1
